+++ include/protocol/websocket.hpp
@@ -2,7 +2,7 @@
  * @file websocket.hpp
  * @author zuudevs (zuudevs@gmail.com)
  * @brief WebSocket protocol support (RFC 6455)
- * @version 0.4.0
+ * @version 1.0.0
  * @date 2026-02-19
  * 
  * @copyright Copyright (c) 2026