+++ include/security/certificate.hpp
@@ -2,7 +2,7 @@
  * @file certificate.hpp
  * @author zuudevs (zuudevs@gmail.com)
  * @brief Certificate information utilities
- * @version 0.5.0
+ * @version 1.0.0
  * @date 2026-02-19
  * 
  * @copyright Copyright (c) 2026